WEBAPP-311: Move layout components to separate packages
