+++ src/modules/app/containers/Switcher.js
@@ -31,8 +31,8 @@ import CityModel from '../../endpoint/models/CityModel'
 import LoadingSpinner from '../../common/components/LoadingSpinner'
 import Layout from '../../layout/components/Layout'
 import LocationLayout, { LocationLayoutRoutes } from '../../layout/containers/LocationLayout'
-import GeneralHeader from '../../layout/components/GeneralHeader'
-import GeneralFooter from '../../layout/components/GeneralFooter'
+import GeneralHeader from '../../layout/components/header/GeneralHeader'
+import GeneralFooter from '../../layout/components/footer/GeneralFooter'
 import type { StateType } from '../StateType'
 import type { Node } from 'react'
 import { SPRUNGBRETT_ROUTE } from '../routes/sprungbrett'
