+++ src/modules/layout/components/footer/__tests__/GeneralFooter.spec.js
