+++ src/modules/layout/components/footer/__tests__/LocationFooter.spec.js
