+++ src/modules/layout/components/header/Header.js
@@ -1,15 +1,15 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import style from './Header.css'
-import logoWide from '../assets/integreat-app-logo.png'
+import logoWide from '../../assets/integreat-app-logo.png'
 import HeaderNavigationBar from './HeaderNavigationBar'
 import HeaderActionBar from './HeaderActionBar'
-import HeaderActionItem from '../HeaderActionItem'
+import HeaderActionItem from '../../HeaderActionItem'
 import Link from 'redux-first-router-link'
-import Headroom from '../../common/components/Headroom'
+import Headroom from '../../../common/components/Headroom'
 import { withTheme } from 'styled-components'
-import withPlatform from '../../platform/hocs/withPlatform'
-import Platform from '../../platform/Platform'
+import withPlatform from '../../../platform/hocs/withPlatform'
+import Platform from '../../../platform/Platform'
 import { compose } from 'redux'
 
 /**
