+++ src/modules/layout/components/header/HeaderNavigationItem.styles.js
