+++ src/modules/layout/components/header/__tests__/HeaderNavigationBar.spec.js
