+++ src/modules/layout/components/toolbar/Toolbar.css
@@ -1,4 +1,4 @@
-@import url('../../app/constants/dimensions.css');
+@import url('../../../app/constants/dimensions.css');
 
 .toolbar {
   display: flex;
